Move utils/date to core/dom/parse-date-attributes
